[TASK] Move a couple of test fixtures and remove obsolete ones
Various fixture files are unused. Remove these. Move
some others around. The "User" classes are still used
in unit tests, but are also moved to functional fixture
folder. This will be resolved later.
