+++ phpstan-baseline.neon
@@ -201,12 +201,12 @@ parameters:
 			path: src/ViewHelpers/SectionViewHelper.php
 
 		-
-			message: "#^Method TYPO3Fluid\\\\Fluid\\\\Tests\\\\Functional\\\\Fixtures\\\\ViewHelpers\\\\ParsedTemplateImplementationFixture\\:\\:getLayoutName\\(\\) should return string but return statement is missing\\.$#"
+			message: "#^Method TYPO3Fluid\\\\Fluid\\\\Tests\\\\Functional\\\\Fixtures\\\\Various\\\\ParsedTemplateImplementationFixture\\:\\:getLayoutName\\(\\) should return string but return statement is missing\\.$#"
 			count: 1
-			path: tests/Functional/Fixtures/ViewHelpers/ParsedTemplateImplementationFixture.php
+			path: tests/Functional/Fixtures/Various/ParsedTemplateImplementationFixture.php
 
 		-
-			message: "#^Call to static method PHPUnit\\\\Framework\\\\Assert\\:\\:assertSame\\(\\) with TYPO3Fluid\\\\Fluid\\\\Tests\\\\Unit\\\\ViewHelpers\\\\Fixtures\\\\UserWithoutToString and string will always evaluate to false\\.$#"
+			message: "#^Call to static method PHPUnit\\\\Framework\\\\Assert\\:\\:assertSame\\(\\) with TYPO3Fluid\\\\Fluid\\\\Tests\\\\Functional\\\\Fixtures\\\\Various\\\\UserWithoutToString and string will always evaluate to false\\.$#"
 			count: 1
 			path: tests/Functional/ViewHelpers/Format/HtmlspecialcharsViewHelperTest.php
