+++ tests/Functional/Fixtures/Various/RenderableFixture.php
@@ -7,7 +7,7 @@ declare(strict_types=1);
  * See LICENSE.txt that was shipped with this package.
  */
 
-namespace TYPO3Fluid\Fluid\Tests\Functional\Fixtures\ViewHelpers;
+namespace TYPO3Fluid\Fluid\Tests\Functional\Fixtures\Various;
 
 use TYPO3Fluid\Fluid\Core\Parser\SyntaxTree\NodeInterface;
 use TYPO3Fluid\Fluid\Core\Parser\SyntaxTree\TextNode;
