+++ tests/Functional/ViewHelpers/RenderViewHelperTest.php
@@ -10,7 +10,7 @@ declare(strict_types=1);
 namespace TYPO3Fluid\Fluid\Tests\Functional\ViewHelpers;
 
 use TYPO3Fluid\Fluid\Tests\Functional\AbstractFunctionalTestCase;
-use TYPO3Fluid\Fluid\Tests\Functional\Fixtures\ViewHelpers\RenderableFixture;
+use TYPO3Fluid\Fluid\Tests\Functional\Fixtures\Various\RenderableFixture;
 use TYPO3Fluid\Fluid\View\Exception\InvalidSectionException;
 use TYPO3Fluid\Fluid\View\Exception\InvalidTemplateResourceException;
 use TYPO3Fluid\Fluid\View\TemplateView;
@@ -75,7 +75,7 @@ class RenderViewHelperTest extends AbstractFunctionalTestCase
     public function renderDataProvider(): \Generator
     {
         yield 'delegate' => [
-            '<f:render delegate="TYPO3Fluid\Fluid\Tests\Functional\Fixtures\ViewHelpers\ParsedTemplateImplementationFixture"/>',
+            '<f:render delegate="TYPO3Fluid\Fluid\Tests\Functional\Fixtures\Various\ParsedTemplateImplementationFixture"/>',
             [],
             'rendered by fixture',
         ];
