--- tests/Unit/ViewHelpers/Fixtures/IfFixture.html
@@ -1,1 +1,0 @@
-{namespace f=TYPO3Fluid\Fluid\ViewHelpers}<f:if condition="{condition}">RenderSomething</f:if>
