--- tests/Unit/ViewHelpers/Fixtures/IfThenElseFixture.html
@@ -1,5 +1,0 @@
-{namespace f=TYPO3Fluid\Fluid\ViewHelpers}<f:if condition="{condition}">
-Do not display
-<f:then>YEP</f:then>
-<f:else>NOPE</f:else>
-</f:if>
